Scored Cognition Data

A simulated dataset with scored cognition test results. This dataset contains calculated total scores for immediate recall, delayed recall, serial subtraction.

Usage

cog_data_score

Format

A dataframe with 10 rows and 6 variables:

HHID

Household identifier, a unique 6-digit integer.

PN

Person number, a unique 1- or 2-digit integer within each household.

Total_I

Total immediate word recall score, ranging from 0 to 5 (sum of 5 items from the immediate recall test).

Total_D

Total delayed word recall score, ranging from 0 to 5 (sum of 5 items from the delayed recall test).

Total_Sub

Total serial subtraction score, ranging from 0 to 5 (sum of successful subtractions from the serial subtraction test).

Total_Count

Total backwards counting score, ranging from 0 to 2 (2 points for success on the first try, 1 point for success on the second try, and 0 for failure).

Examples

# Load the data
data(cog_data_score)

# View the first few rows
head(cog_data_score)
#>     HHID PN Total_I Total_D Total_Sub Total_Count
#> 1 288941 93       9       7         0           0
#> 2 234057 99       6       7         0           1
#> 3 224021 72       8       8         0           0
#> 4 785284 26       6       3         1           0
#> 5 326317  7       8       7         0           2
#> 6 465208 42       7       6         0           2